app/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "API for Bloom platform"
authors = [
    { name = "Bloom Team", email = "dev@beetletld.com" }
]
license = { file = "../../../LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "babel>=2.17.0",
    "boto3>=1.40.1",
    "celery-types>=0.23.0",
    "celery[redis]>=5.5.3",
    "cryptography>=45.0.6",
    "fastapi-problem>=0.11.4",
    "fastapi[standard]>=0.116.1",
    "inflection>=0.5.1",
    "iso4217>=1.14.20250512",
    "limits>=5.5.0",
    "mjml-python>=1.3.5",
    "passlib>=1.7.4",
    "phonenumbers>=9.0.10",
    "psycopg[binary]>=3.2.9",
    "pycountries>=1.2.1",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.42",
    "sqlmodel>=0.0.24",
    "tenacity>=9.1.2",
]

[dependency-groups]
dev = [
    "coverage>=7.10.2",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "types-passlib>=1.7.7.20250602",
]
